translation/requirements.txt
google-cloud-translate
google-cloud-pubsub
flask
